file.pyQRCode/requirements.txt
pyqrcode
python-cv2
